decoupling annotation and attribute drivers
