+++ .phpunit.result.cache
@@ -1,0 +1,1 @@
+{"version":1,"defects":[],"times":{"Hateoas\\Tests\\Configuration\\Metadata\\Driver\\AttributeDriverTest::testUser":0.03,"Hateoas\\Tests\\Configuration\\Metadata\\Driver\\AttributeDriverTest::testEmptyClass":0.001,"Hateoas\\Tests\\Configuration\\Metadata\\Driver\\AnnotationDriverTest::testUser":0.048,"Hateoas\\Tests\\Configuration\\Metadata\\Driver\\AnnotationDriverTest::testEmptyClass":0.001}}
